Added GetLastAUGShotNumber which returns the current Shotnumber of ASDEX Upgrade.
Added GetLastShotNumber(Experiment, Diagnostic, PulseNumber=None) which
returns the LastShotNumber of the given Experiment and Diagnostic. If
a PulseNumber is given then the LastShotNumber starting form this
PulseNumber is returned. If the shotfile is available for this
PulseNumber the result will be PulseNumber. If no PulseNumber is
specified then the last PulseNumber of ASDEX Upgrade will be used as an
input (from GetLastAUGShotNumber).