Added getSignal(name, dtype=None)
If no dtype is specified the data is returned using the datatype
determined from 'dataformat'. If dtype is specified the data is
converted accordingly.

So far I tested it on the DCN and it works with SHORT INTEGERS (int16).
The string arrays need to be implemented, see therefore __dataformat__.
If someone could provide a diagnostic where string signals are used I
would be happy to test it.