Added check if signalGroup has proper shape.
In case the time is not the first index in the signal group the (e.g. in
IDA) the usage of tBegin and tEnd is disabled. For such signals libddww
does not support reading time slices out of the box. getSignalGroup was
now tested with CEC and IDA and it works properly for both diagnostics.