Changed warnings and getSignal / Calibrated
Changed warning system to module warnings. These do not interrupt code
execution and they can be suppressed.

getSignal and getSignalCalibrated can now read signals that have no
timebase associated with them. Example: PDP -> LOCATION.
The getTimeBaseInfo throws an exception in this case which is used to
set k1 = 1 and k2 = index[0] of the signalInfo.