Added getSignalGroupCalibrated
getSignalGroupCalibrated returns the calibrated data of the signalGroup.
__call__ uses getSignalGroupCalibrated to return the calibrated data of
signalGroup. The keywords tBegin, tEnd and dtype have the same behavior
as in getSignalCalibrated. Since the calibrated data can only be
converted to float32 of float64 a check is performed if the provided
dtype is valid. If it is not valid numpy.float32 is used as dtype.