Added getParameter and parameter
Added getParameter which returns the specified parameter of the
specified parameter set as a parameter object. The object contains the
set name, parameter name, data and the human readable units.

getParameter takes a dtype if no dtype is specified the RAW data from
the shotfile should be read. However this does not work. libddww returns
an empty buffer.

Note the datatype 1974 has been added to __dataformat__ returning
numpy.dtype('S8') which is the corresponding data format for CHAR*8.
Unfortunately this could not yet be tested due to the bug mentioned
above.