Added signal class
The signal class is intended to be returned by __call__ if the object is
a signal. It contains the signal name, data and if available the time
and the unit.

For the class basic functionality has been implemented.

__call__(tBegin, tEnd) returns a signal class containing the signal in
the specified time interval.

max(), min() and median() have been implemented. In case of min and max
NaNs are ignored.

The Idea is:

sf = dd.shotfile('DCN', 29708)
H1 = a('H-1')
medianValue = H2(1.5, 1.6).median()