Added getObjectData.
getObjectData returns the RAW data of the object in the appropriate data
format. In case of signalGroups and AreaBases the shape of the output
array is changed to the shape stored in the objectHeader.

Checks need to be made if the formatting works for different
diagnostics. So far IDA and PDP SignalGroups and AreaBases are tested.