Revised getAreaBase.
getAreaBase now uses getObjectData to load the areabase.
If a signal / signal group is specified all area bases will be loaded
and returned in a dictionary.
If only one area base exists the area base is returned.

The area bases are contained in the areaBase class containing the name
and the data.