Fixed error in dd.shotfile.open
For some discharges it was not possible to open specific shotfiles if
other discharges of the same diagnostic were opened before.

This was the result of an error in ddlastshotnr_ which was used to check
if the shotfile for the desired discharge is available. This check has
been removed. If the shotfile is not available an exception is thrown.